Add embeddable ShareButton for resource-sharing consumer plugins
Exposes securityDashboards.ui.ShareButton on the plugin start contract so
plugins that register protected resources under the resource-sharing
framework can embed a Share / Update Access button (with the shared
access modal) directly on their own pages, instead of navigating to the
central Resource Access Management dashboard.

- Extract ShareAccessModal, share-with types and diff/error helpers out
  of resource-sharing-panel into reusable modules (no behavior change)
- Add ResourceShareButton: self-fetches sharing record + access levels,
  handles can_share permission, MDS dataSourceId, and hides itself when
  resource sharing is disabled or the type is not registered
- Lazy-load the implementation so consumers only pay the cost on render
