+++ public/apps/resource-sharing/resource-sharing-panel.tsx
@@ -20,19 +20,9 @@ import {
   EuiBadge,
   EuiButton,
   EuiButtonEmpty,
-  EuiCallOut,
-  EuiComboBox,
   EuiFlexGroup,
   EuiFlexItem,
-  EuiForm,
-  EuiFormRow,
-  EuiModal,
-  EuiModalBody,
-  EuiModalFooter,
-  EuiModalHeader,
-  EuiModalHeaderTitle,
   EuiPanel,
-  EuiSpacer,
   EuiText,
   EuiToolTip,
   EuiSuperSelect,
@@ -43,449 +33,14 @@ import _ from 'lodash';
 
 import type { CoreStart } from '../../../../../src/core/public';
 
-interface CreatedBy {
-  user: string;
-  tenant?: string;
-}
-
-interface ShareRecipients {
-  users?: string[];
-  roles?: string[];
-  backend_roles?: string[];
-}
-
-type ShareWith = Record<string /* access-level e.g. READ_ONLY */, ShareRecipients>;
-
-interface ResourceRow {
-  resource_id: string;
-  resource_type: string;
-  created_by: CreatedBy;
-  share_with?: ShareWith; // may be empty/undefined
-  can_share?: boolean; // whether the current user can share this resource
-}
-interface TypeEntry {
-  type: string; // type of resource, e.g. `sample-resource`
-  access_levels: string[]; // known access-levels for this type
-}
-
-// API that the panel consumes
-interface Api {
-  listTypes: () => Promise<{ types: TypeEntry[] } | TypeEntry[]>;
-  listSharingRecords: (type: string) => Promise<ResourceRow[] | { resources: ResourceRow[] } | any>;
-  getSharingRecord: (
-    id: string,
-    type: string
-  ) => Promise<ResourceRow | { resource: ResourceRow } | any>;
-  share: (payload: {
-    resource_id: string;
-    resource_type: string;
-    share_with: ShareWith;
-  }) => Promise<any>;
-  update: (payload: {
-    resource_id: string;
-    resource_type: string;
-    add?: ShareWith;
-    revoke?: ShareWith;
-  }) => Promise<any>;
-}
+import { ResourceRow, ResourceSharingApi, ShareWith, TypeEntry } from './types';
+import { hasSharingInfo } from './share-utils';
+import { ShareAccessModal } from './share-access-modal';
 
 interface Props {
-  api: Api;
+  api: ResourceSharingApi;
   toasts: CoreStart['notifications']['toasts'];
 }
-
-const hasSharingInfo = (sw?: ShareWith) =>
-  !!sw &&
-  Object.values(sw).some(
-    (v) => !!v && (v.users?.length || v.roles?.length || v.backend_roles?.length)
-  );
-
-/** ---------- helpers: UI <-> payload ---------- */
-const toOptions = (vals?: string[]) => (vals || []).map((v) => ({ label: v }));
-const fromOptions = (opts: Array<{ label: string }>) => opts.map((o) => o.label);
-const cloneShareWith = (sw?: ShareWith): ShareWith => JSON.parse(JSON.stringify(sw || {}));
-
-/** diff: produce { add, revoke } between old and next (both are ShareWith): for PUT/PATCH api calls */
-function diffShareWith(prev: ShareWith, next: ShareWith): { add?: ShareWith; revoke?: ShareWith } {
-  const add: ShareWith = {};
-  const revoke: ShareWith = {};
-  const allLevels = new Set<string>([...Object.keys(prev || {}), ...Object.keys(next || {})]);
-
-  for (const g of allLevels) {
-    const p = prev?.[g] || {};
-    const n = next?.[g] || {};
-    const keys: Array<keyof ShareRecipients> = ['users', 'roles', 'backend_roles'];
-
-    for (const k of keys) {
-      const pSet = new Set(p[k] || []);
-      const nSet = new Set(n[k] || []);
-
-      const adds = [...nSet].filter((x) => !pSet.has(x));
-      if (adds.length) {
-        add[g] = add[g] || {};
-        (add[g] as any)[k] = [...((add[g] as any)[k] || []), ...adds];
-      }
-
-      const removals = [...pSet].filter((x) => !nSet.has(x));
-      if (removals.length) {
-        revoke[g] = revoke[g] || {};
-        (revoke[g] as any)[k] = [...((revoke[g] as any)[k] || []), ...removals];
-      }
-    }
-  }
-
-  const empty = (sw: ShareWith) =>
-    !Object.keys(sw).length ||
-    Object.values(sw).every(
-      (v) => !v.users?.length && !v.roles?.length && !v.backend_roles?.length
-    );
-
-  return { add: empty(add) ? undefined : add, revoke: empty(revoke) ? undefined : revoke };
-}
-
-/** ----- Share Modal error helper ------ */
-const parseMaybeJson = (s: unknown) => {
-  if (typeof s !== 'string') return undefined;
-  try {
-    return JSON.parse(s);
-  } catch {
-    return undefined;
-  }
-};
-
-const extractHttpErrorLines = (e: any): string[] => {
-  const lines: string[] = [];
-
-  // 1) Basic status / message
-  if (e?.response?.status || e?.response?.statusText) {
-    lines.push(`${e?.response?.status ?? ''} ${e?.response?.statusText ?? ''}`.trim());
-  }
-  if (e?.message && e.message !== e?.response?.statusText) lines.push(e.message);
-
-  // 2) Known body shapes
-  let body: any = e?.body;
-  if (typeof body === 'string') body = parseMaybeJson(body) ?? { raw: body };
-
-  // OSD style Error: { statusCode, error, message }
-  if (body?.statusCode || body?.error || body?.message) {
-    if (body?.error && typeof body.error === 'string') lines.push(body.error);
-    if (body?.message && typeof body.message === 'string') lines.push(body.message);
-  }
-
-  // OpenSearch style Error: { error: { reason, root_cause[], caused_by }, status }
-  const osErr = body?.error;
-  if (osErr) {
-    if (typeof osErr.reason === 'string') lines.push(osErr.reason);
-    if (osErr.caused_by?.reason) lines.push(osErr.caused_by.reason);
-    if (Array.isArray(osErr.root_cause)) {
-      for (const rc of osErr.root_cause) {
-        if (rc?.reason) lines.push(rc.reason);
-      }
-    }
-    if (Array.isArray(osErr.caused_by?.root_cause)) {
-      for (const rc of osErr.caused_by.root_cause) {
-        if (rc?.reason) lines.push(rc.reason);
-      }
-    }
-  }
-
-  // Raw fallback if we only had a string body
-  if (body?.raw && typeof body.raw === 'string') {
-    lines.push(body.raw);
-  }
-
-  // Dedup & clean
-  return [...new Set(lines.filter(Boolean).map((s) => String(s).trim()))].slice(0, 10);
-};
-
-/** ---------- Share/Update modal ---------- */
-
-interface ModalProps {
-  mode: 'create' | 'edit';
-  isOpen: boolean;
-  onClose: () => void;
-  onSubmit: (payload: any) => Promise<void>;
-  resource: ResourceRow;
-  resourceType: string;
-  resourceTypeIndex: string;
-  accessLevels: string[];
-}
-
-const hasNonEmptyRecipients = (r?: ShareRecipients) =>
-  !!r && !!((r.users?.length ?? 0) || (r.roles?.length ?? 0) || (r.backend_roles?.length ?? 0));
-
-const hasNonEmptyShareWith = (sw?: ShareWith) =>
-  Object.values(sw || {}).some(hasNonEmptyRecipients);
-
-const emptyLevels = (sw?: ShareWith) =>
-  Object.entries(sw || {})
-    .filter(([, r]) => !hasNonEmptyRecipients(r))
-    .map(([g]) => g);
-
-const ShareAccessModal: React.FC<ModalProps> = ({
-  mode,
-  isOpen,
-  onClose,
-  onSubmit,
-  resource,
-  resourceType,
-  resourceTypeIndex,
-  accessLevels,
-}) => {
-  const original = useMemo(() => cloneShareWith(resource?.share_with), [resource?.share_with]);
-  const [working, setWorking] = useState<ShareWith>(() =>
-    mode === 'edit' ? cloneShareWith(resource.share_with) : {}
-  );
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [errorLines, setErrorLines] = useState<string[]>([]);
-
-  useEffect(() => {
-    if (mode === 'edit') setWorking(cloneShareWith(resource.share_with));
-    else setWorking({});
-    setErrorLines([]);
-  }, [mode, resource]);
-
-  const levels = Object.keys(working);
-  // suggestions list for this modal (unique + keep order)
-  const ACCESS_LEVELS = useMemo(() => Array.from(new Set(accessLevels ?? [])), [accessLevels]);
-
-  const addLevel = () => {
-    // If no suggestions were provided for this type, do nothing.
-    if (!ACCESS_LEVELS.length) {
-      return;
-    }
-
-    // Pick the first unused suggestion
-    const base = ACCESS_LEVELS.find((level) => !levels.includes(level));
-    if (!base) return; // all suggestions already used
-
-    setWorking({ ...working, [base]: {} });
-  };
-
-  const removeLevel = (g: string) => {
-    const next = cloneShareWith(working);
-    delete next[g];
-    setWorking(next);
-  };
-
-  const setLevelName = (g: string, newName: string) => {
-    if (!newName || newName === g) return;
-    if (working[newName]) return;
-    const copy = cloneShareWith(working);
-    copy[newName] = copy[g] || {};
-    delete copy[g];
-    setWorking(copy);
-  };
-
-  const setRecipients = (g: string, key: keyof ShareRecipients, values: string[]) => {
-    setWorking((prev) => ({
-      ...prev,
-      [g]: {
-        ...(prev[g] || {}),
-        [key]: values,
-      },
-    }));
-  };
-
-  // compute whether there are changes
-  const diff = useMemo(() => diffShareWith(original || {}, working || {}), [original, working]);
-  const levelsWithNoRecipients = useMemo(() => emptyLevels(working), [working]);
-  const hasChanges =
-    mode === 'create' ? hasNonEmptyShareWith(working) : Boolean(diff.add || diff.revoke);
-  const isInvalid = levelsWithNoRecipients.length > 0;
-
-  const disabledReason = (): string | undefined => {
-    if (isInvalid) {
-      return `These access-levels have no recipients: ${levelsWithNoRecipients.join(', ')}`;
-    }
-
-    if (!hasChanges) {
-      return mode === 'create'
-        ? 'Add at least one user, role, or backend role to share.'
-        : 'No changes detected.';
-    }
-
-    return undefined;
-  };
-
-  const handleSubmit = async () => {
-    setIsSubmitting(true);
-    setErrorLines([]);
-    try {
-      if (mode === 'create') {
-        await onSubmit({
-          resource_id: resource.resource_id,
-          resource_type: resourceTypeIndex,
-          share_with: working,
-        });
-      } else {
-        if (!diff.add && !diff.revoke) {
-          // Shouldn’t happen since button is hidden, but safe-guard anyway
-          setIsSubmitting(false);
-          return;
-        }
-        const payload: any = {
-          resource_id: resource.resource_id,
-          resource_type: resourceTypeIndex,
-          ...(diff.add ? { add: diff.add } : {}),
-          ...(diff.revoke ? { revoke: diff.revoke } : {}),
-        };
-        await onSubmit(payload);
-      }
-      onClose();
-    } catch (e: any) {
-      const lines = extractHttpErrorLines(e);
-      setErrorLines(lines.length ? lines : ['Failed to submit changes.']);
-    } finally {
-      setIsSubmitting(false);
-    }
-  };
-
-  if (!isOpen) return null;
-
-  return (
-    <EuiModal onClose={onClose} style={{ width: 700 }}>
-      <EuiModalHeader>
-        <EuiModalHeaderTitle>
-          {mode === 'create' ? 'Share Resource' : 'Update Access'}
-        </EuiModalHeaderTitle>
-      </EuiModalHeader>
-      <EuiModalBody>
-        <EuiText size="s">
-          <p>
-            <strong>Resource:</strong> {resource.resource_id} &nbsp;·&nbsp; <strong>Type:</strong>{' '}
-            {resourceType}
-          </p>
-        </EuiText>
-        <EuiSpacer size="m" />
-        {!isSubmitting && errorLines.length > 0 && (
-          <>
-            <EuiCallOut announceOnMount title="Request failed" color="danger" iconType="alert">
-              <ul style={{ margin: 0, paddingLeft: 18 }}>
-                {errorLines.map((l, i) => (
-                  <li key={i}>{l}</li>
-                ))}
-              </ul>
-            </EuiCallOut>
-            <EuiSpacer />
-          </>
-        )}
-        {isInvalid && (
-          <>
-            <EuiCallOut announceOnMount title="Add recipients" color="warning" iconType="alert">
-              The following access-levels have no recipients: {levelsWithNoRecipients.join(', ')}
-            </EuiCallOut>
-            <EuiSpacer size="s" />
-          </>
-        )}
-        <EuiForm component="form">
-          <EuiFormRow label="Access-levels">
-            <EuiButtonEmpty iconType="plusInCircle" onClick={addLevel}>
-              Add access-level
-            </EuiButtonEmpty>
-          </EuiFormRow>
-
-          {Object.keys(working).length === 0 && (
-            <EuiText color="subdued" size="s">
-              No access-levels added yet.
-            </EuiText>
-          )}
-
-          {Object.entries(working).map(([levelName, recipients]) => {
-            const levelOptions = [...new Set([levelName, ...ACCESS_LEVELS])].map((l) => ({
-              label: l,
-            }));
-
-            return (
-              <EuiPanel key={levelName} paddingSize="m" hasShadow={false} hasBorder>
-                <EuiFlexGroup gutterSize="m" alignItems="center">
-                  <EuiFlexItem grow={3}>
-                    <EuiFormRow label="Access-level">
-                      <EuiComboBox
-                        singleSelection={{ asPlainText: true }}
-                        options={levelOptions}
-                        selectedOptions={[{ label: levelName }]}
-                        onChange={(opts) => {
-                          const newLabel = opts[0]?.label || levelName;
-                          setLevelName(levelName, newLabel);
-                        }}
-                        onCreateOption={(label: string) => setLevelName(levelName, label)}
-                      />
-                    </EuiFormRow>
-                  </EuiFlexItem>
-                  <EuiFlexItem grow={false}>
-                    <EuiButtonEmpty
-                      color="danger"
-                      iconType="trash"
-                      onClick={() => removeLevel(levelName)}
-                    >
-                      Remove
-                    </EuiButtonEmpty>
-                  </EuiFlexItem>
-                </EuiFlexGroup>
-
-                <EuiFormRow label="Users">
-                  <EuiComboBox
-                    placeholder="Add users…"
-                    noSuggestions
-                    selectedOptions={toOptions(recipients.users)}
-                    onCreateOption={(v) =>
-                      setRecipients(levelName, 'users', [...(recipients.users || []), v])
-                    }
-                    onChange={(opts) => setRecipients(levelName, 'users', fromOptions(opts))}
-                  />
-                </EuiFormRow>
-                <EuiFormRow label="Roles">
-                  <EuiComboBox
-                    placeholder="Add roles…"
-                    noSuggestions
-                    selectedOptions={toOptions(recipients.roles)}
-                    onCreateOption={(v) =>
-                      setRecipients(levelName, 'roles', [...(recipients.roles || []), v])
-                    }
-                    onChange={(opts) => setRecipients(levelName, 'roles', fromOptions(opts))}
-                  />
-                </EuiFormRow>
-                <EuiFormRow label="Backend roles">
-                  <EuiComboBox
-                    placeholder="Add backend roles…"
-                    noSuggestions
-                    selectedOptions={toOptions(recipients.backend_roles)}
-                    onCreateOption={(v) =>
-                      setRecipients(levelName, 'backend_roles', [
-                        ...(recipients.backend_roles || []),
-                        v,
-                      ])
-                    }
-                    onChange={(opts) =>
-                      setRecipients(levelName, 'backend_roles', fromOptions(opts))
-                    }
-                  />
-                </EuiFormRow>
-              </EuiPanel>
-            );
-          })}
-        </EuiForm>
-      </EuiModalBody>
-      <EuiModalFooter>
-        <EuiButtonEmpty onClick={onClose} isDisabled={isSubmitting}>
-          Cancel
-        </EuiButtonEmpty>
-        <EuiToolTip content={disabledReason()}>
-          <EuiButton
-            onClick={handleSubmit}
-            fill
-            isLoading={isSubmitting}
-            isDisabled={!hasChanges || isInvalid || isSubmitting}
-          >
-            {mode === 'create' ? 'Share' : 'Update Access'}
-          </EuiButton>
-        </EuiToolTip>
-      </EuiModalFooter>
-    </EuiModal>
-  );
-};
-
 /** ---------- Expanded view for Shared With ---------- */
 const SharedWithExpanded: React.FC<{ sw?: ShareWith }> = ({ sw }) => {
   if (!hasSharingInfo(sw)) {
